src/test/java/com/springboot/project/web/HelloControllerTest.java
package com.springboot.project.web;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.is;
import org.springframework.test.web.servlet.ResultActions;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)  // 1
@WebMvcTest(controllers = HelloController.class) // 2
public class HelloControllerTest {
    @Autowired  // 3
    private MockMvc mvc;  // 4

    @Test
    public void hello가_리턴된다() throws Exception{
        String hello = "hello";

        mvc.perform(get("/hello"))  // 5
                .andExpect(status().isOk()) // 6
                .andExpect(content().string(hello));  //7
    }

    @Test
    public void helloDto가_리턴된다() throws Exception{
        String name = "hello";
        int amount = 1000;

        mvc.perform(
                get("/hello/dto")
                        .param("name",name) //API테스트할 때 사용될 요청 파라미터 설정, 단 String만 가능!
                        .param("amount",String.valueOf(amount)))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.name",is(name))) //JSON응답값을 필드별로 검증할 수 있음, $를 기준으로 필드명 명시
                .andExpect(jsonPath("$.amount",is(amount)));
    }
}